refactor: split rust auth crate into focused modules
