+++ tests/loadtest/run_core_auth_compare.sh
