+++ tools_rust/auth/src/config.rs
@@ -24,19 +24,35 @@ pub struct AuthConfig {
     #[arg(long, env = "CONTEXTFORGE_AUTH_BACKEND_HEALTH_URL")]
     pub backend_health_url: Option<String>,
 
-    #[arg(long, env = "CONTEXTFORGE_AUTH_LISTEN_HTTP", default_value = "127.0.0.1:8788")]
+    #[arg(
+        long,
+        env = "CONTEXTFORGE_AUTH_LISTEN_HTTP",
+        default_value = "127.0.0.1:8788"
+    )]
     pub listen_http: String,
 
-    #[arg(long, env = "CONTEXTFORGE_AUTH_REQUEST_TIMEOUT_MS", default_value_t = 30_000)]
+    #[arg(
+        long,
+        env = "CONTEXTFORGE_AUTH_REQUEST_TIMEOUT_MS",
+        default_value_t = 30_000
+    )]
     pub request_timeout_ms: u64,
 
     #[arg(long, env = "CONTEXTFORGE_AUTH_DB_POOL_MAX_SIZE", default_value_t = 8)]
     pub db_pool_max_size: usize,
 
-    #[arg(long, env = "CONTEXTFORGE_AUTH_EXPERIMENTAL_DIRECT_AUTH", default_value_t = false)]
+    #[arg(
+        long,
+        env = "CONTEXTFORGE_AUTH_EXPERIMENTAL_DIRECT_AUTH",
+        default_value_t = false
+    )]
     pub experimental_direct_auth: bool,
 
-    #[arg(long, env = "CONTEXTFORGE_AUTH_SHADOW_COMPARE_DIRECT_AUTH", default_value_t = false)]
+    #[arg(
+        long,
+        env = "CONTEXTFORGE_AUTH_SHADOW_COMPARE_DIRECT_AUTH",
+        default_value_t = false
+    )]
     pub shadow_compare_direct_auth: bool,
 
     #[arg(long, env = "CONTEXTFORGE_AUTH_LOG", default_value = "info")]
@@ -69,14 +85,21 @@ pub struct AuthConfig {
     #[arg(long, env = "REQUIRE_USER_IN_DB", default_value_t = false)]
     pub require_user_in_db: bool,
 
-    #[arg(long, env = "PLATFORM_ADMIN_EMAIL", default_value = "admin@example.com")]
+    #[arg(
+        long,
+        env = "PLATFORM_ADMIN_EMAIL",
+        default_value = "admin@example.com"
+    )]
     pub platform_admin_email: String,
 }
 
 impl AuthConfig {
     pub fn listen_addr(&self) -> Result<SocketAddr, String> {
-        self.listen_http
-            .parse()
-            .map_err(|err| format!("invalid CONTEXTFORGE_AUTH_LISTEN_HTTP '{}': {err}", self.listen_http))
+        self.listen_http.parse().map_err(|err| {
+            format!(
+                "invalid CONTEXTFORGE_AUTH_LISTEN_HTTP '{}': {err}",
+                self.listen_http
+            )
+        })
     }
 }
