+++ tools_rust/auth/src/main.rs
@@ -12,7 +12,9 @@ async fn main() {
     let config = AuthConfig::parse();
 
     tracing_subscriber::fmt()
-        .with_env_filter(tracing_subscriber::EnvFilter::new(config.log_filter.clone()))
+        .with_env_filter(tracing_subscriber::EnvFilter::new(
+            config.log_filter.clone(),
+        ))
         .init();
 
     if let Err(err) = run(config).await {
